requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0
streamlit==1.27.0
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.3.0
tqdm==4.65.0